JUS-742 Updates the dynamic table filter modal position calculation
The getPosition function calculated the position based on the parent
positioned element. (In this case with respect to the table)
Originally the table head always appeared on the top of the table,
so the difference was not apparent.
Now the table head is fixed to the top of the container and the table
scrolls under the head. This results in different position values when
calculated wrt the th and the table.

This patch uses jQuery offset method to calculate the position wrt
the th alone.